Use int64_t instead of long long and LL

+++ lib/template.cpp
@@ -163,10 +163,10 @@ struct edge {
     }
 };
 
-using LL = long long;
+using LL = int64_t;
 
 #define fs first
 #define sc second
 
-const LL MOD = 1e9+7;
+const int64_t MOD = 1e9+7;
 #endif
